Support pure go sqlite driver
