+++ go.mod
@@ -23,4 +23,5 @@ require (
 	go.etcd.io/etcd/client/v3 v3.5.0
 	go.etcd.io/etcd/server/v3 v3.5.0
 	google.golang.org/grpc v1.38.0
+	modernc.org/sqlite v1.19.1
 )
